+++ functions/database-debug.log
@@ -1,0 +1,13 @@
+13:20:16.696 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO akka.event.slf4j.Slf4jLogger - Slf4jLogger started
+13:20:16.791 [main] INFO com.firebase.server.forge.App$ - Listening at 127.0.0.1:9000
+13:20:28.717 [NamespaceSystem-akka.actor.default-dispatcher-5] INFO com.firebase.core.namespace.NamespaceActor - phoenix-370-default-rtdb successfully activated FBKV (SurveyIdle(0)) wait: 76ms, init: 0ms
+13:21:04.755 [FirebaseWorkerPool-1-3] WARN com.firebase.core.persistence.backend.dummy.EmulatorMetadataPersistence - Multiple projectIds are not recommended in single project mode. Requested 
+namespace phoenix-370, but the emulator is configured for phoenix-370-default-rtdb. To 
+opt-out of single project mode add/set the single_project_mode: false property in 
+the firebase.json emulators config. 
+13:21:04.765 [NamespaceSystem-akka.actor.default-dispatcher-7] INFO com.firebase.core.namespace.NamespaceActor - phoenix-370 successfully activated FBKV (SurveyIdle(0)) wait: 1ms, init: 0ms
+13:22:59.605 [Thread-0] INFO com.firebase.server.forge.App$ - Attempting graceful shutdown.
+13:22:59.611 [NamespaceSystem-akka.actor.default-dispatcher-31] INFO com.firebase.core.namespace.Terminator$Terminator - 2 actors left to terminate: phoenix-370-default-rtdb, phoenix-370
+13:22:59.613 [NamespaceSystem-akka.actor.default-dispatcher-31] INFO com.firebase.core.namespace.NamespaceActor - stopped namespace actor for phoenix-370
+13:22:59.613 [NamespaceSystem-akka.actor.default-dispatcher-30] INFO com.firebase.core.namespace.NamespaceActor - stopped namespace actor for phoenix-370-default-rtdb
+13:22:59.615 [Thread-0] INFO com.firebase.server.forge.App$ - Graceful shutdown complete.
