+++ functions/package.json
@@ -4,7 +4,7 @@
         "lint": "eslint --ext .js,.ts .",
         "build": "tsc",
         "build:watch": "tsc --watch",
-        "serve": "npm run build && firebase emulators:start --only functions",
+        "serve": "npm run build && firebase emulators:start",
         "shell": "npm run build && firebase functions:shell",
         "start": "npm run shell",
         "deploy": "firebase deploy --only functions",
