+++ functions/pubsub-debug.log
@@ -1,0 +1,11 @@
+This is the Google Pub/Sub fake.
+Implementation may be incomplete or differ from the real system.
+Apr 04, 2023 1:20:16 PM com.google.cloud.pubsub.testing.v1.Main main
+INFO: IAM integration is disabled. IAM policy methods and ACL checks are not supported
+SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".
+SLF4J: Defaulting to no-operation (NOP) logger implementation
+SLF4J: See http://www.slf4j.org/codes.html#StaticLoggerBinder for further details.
+Apr 04, 2023 1:20:17 PM com.google.cloud.pubsub.testing.v1.Main main
+INFO: Server started, listening on 8085
+Apr 04, 2023 1:20:18 PM io.gapi.emulators.netty.HttpVersionRoutingHandler channelRead
+INFO: Detected HTTP/2 connection.
