+++ functions/src/utils.ts
@@ -1,0 +1,9 @@
+import { Clean } from "./types";
+
+export function cleanUndefined<T extends object>(obj: T): T {
+    let clean: Clean<T> = {};
+    for (const k in obj) {
+        if (obj[k]) clean[k] = obj[k];
+    }
+    return clean as T;
+}
